third_party/WebKit/LayoutTests/http/tests/notifications/notification-creation-detached-context.html

<!doctype html>
<html>
  <head>
    <title>Notifications: Creating a notification in a detached context should
      throw a TypeError.</title>
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
    <script src="resources/test-helpers.js"></script>
  </head>
  <body>
    <script>
      if (window.testRunner) {
          testRunner.setPermission('notifications', 'granted', location.origin, location.origin);
          testRunner.setCanOpenWindows();
      }

      async_test(test => {
          const remoteWindow = window.open('resources/window-detached-context.html');
          let notificationObj = null;

          window.addEventListener('message', test.step_func(event => {
            switch (event.data) {
              case 'opened':
                  assert_own_property(remoteWindow, 'Notification');
                  notificationObj = remoteWindow.Notification;
                  remoteWindow.close();
                  break;

              case 'closed':
                  assert_throws(new TypeError(), function() {
                      new notificationObj('Title');
                  }, 'Creating a notification from a detached context should throw a Type Error');

                  test.done();
                  break;

              default:
                  assert_unreached('Unrecognized message from the opened window.');
                  break;
            }
          }));

      }, 'Creating a notification in a detached context should throw.');
    </script>
  </body>
</html>
